+++ monaco/javascript/test/dummy.ts
@@ -1,0 +1,5 @@
+describe('@cucumber/monaco', () => {
+  it('does not have any unit tests', () => {
+    console.log('******* See README.md for details about how to manually test @cucumber/monaco *******')
+  })
+})