+++ suggest/javascript/src/index/fuseIndex.ts
@@ -1,11 +1,12 @@
-import { Index, StepDocument } from '../src'
+import { Index } from './types'
+import { StepDocument } from '../types'
 import Fuse from 'fuse.js'
 
 type Doc = {
   text: string
 }
 
-export default function fuseIndex(stepDocuments: readonly StepDocument[]): Index {
+export function fuseIndex(stepDocuments: readonly StepDocument[]): Index {
   const docs: Doc[] = stepDocuments.map((stepDocument) => {
     return {
       text: stepDocument.segments
@@ -13,9 +14,15 @@ export default function fuseIndex(stepDocuments: readonly StepDocument[]): Index
         .join(''),
     }
   })
-  const fuse = new Fuse(docs, { keys: ['text'], minMatchCharLength: 2, threshold: 0.1, ignoreLocation: true })
+  const fuse = new Fuse(docs, {
+    keys: ['text'],
+    minMatchCharLength: 2,
+    threshold: 0.1,
+    ignoreLocation: true,
+  })
 
   return (text) => {
+    if (!text) return []
     const results = fuse.search(text, { limit: 10 })
     return results.map((result) => stepDocuments[result.refIndex])
   }